Correctly implement diff.
Now that we no longer have recursive dependencies, the `run` function is
my simplier.

At some point we stopped checking out the specific version in the
repository so our initial evaluation of modules was confused.

Passing the `o_barnyard` and `o_module` using environment variables
instead of writing to a temporary file.

Fixed the `barnyard_get_array` to leave array empty if the associative
array entry is an empty string. It was adding a single empty string.