Allow for `search` and `functions` commands.
Simplifications. Development argument is the only real argument now in
addition to module names. Running status functions provided by modules
as if they were module applys. Created a function to emit the common
functions for the user to source should they need them, but they are now
getting included in status since it is invoked as if it were an apply.