Add ability to always other than `apply.zsh`.
Had a command `exec` that I used to execute a program with Barnyard
functions but with no other context. Have `always` which does provide a
clone of a version or runs in a working directory. Updated `always` to
allow you to specify an alternative to `apply.zsh` with
`module/other.zsh`. Now we can run timer programs or maintainence
programs using `always` and have an environment that the program is
familiar with.