Remove `exec`, `status`, and `functions`.
Remove these commands that were an attempt to provide a way to run
Barnyard modules as individual programs. We have that now with `always`
and with `always foo/alt.zsh` where `foo` is the module and `alt.zsh` is
an alternative to `apply.zsh`. We have now cleared out some complexity
and we can work on making `barnyard apply` and `barnyard always` run
faster.