Single module loop function.
Merge the loop modules and run modules functions into one function, but
we have to keep an ordering function separate for now. We need the loop
in a single function so we can export it to the user module so we can
have a context aware `barnyard_always` function for the user that will
run modules from within a Barnyard module without forking Barnyard and
creating a new clone of the repository.

We are going to pull it all into a single function, though, and control
everything with our `o_registers` flags.