Revamped argument parser.
Short arguments actually work. Have negated arguments working for both
long and short arguments. Created a little readme, but it is already out
of date, going to work with the argument parser before I commit to it.

Ability to declare variables works. Now have a way of specifying
defaultsall argument accepting switches just by testing to see if the
value is already defined before redelcaring it.

Added a way to specify values to associative arrays.