Emit `logfmt`.
There's a log format called `logfmt` that is key value pairs with values
in quotes with non-Microsoft-style CSV escaping. Grafana supports
parting it and it is encouraged by Splunk.

At the outset of logging, we attempted to use systemd-journal priority
levels and sought to expand this to use systemd-journal properties going
so far as to attempt to create a unique message id for a message so we
could track it as its line number changes.

That's not the way things work in practice. We want somewhat structured
logging in our database so that we can generate alerts from it, and we
need to be careful not to break the alerts. That's about it.

That's the initial application of Barnyard logging, not instrumentation.

We may learn more about what we need to know as we add instrumentation,
but it is still probably going to be short log lines, not stack traces
or other complicated dumps.