Remove `ls-remote`, fetchless always.
`barnyard always` will not update the mirror by default so we can use
`barnyard always` for utilities that might be on a timer. We do now
already, so we know that we're able to pull from GitHub quite a bit and
not run into problems.

We were using `git ls-remote` but we don't have to when we have a full
local mirror.

Little utilities run in about 200ms now.