Begin using common Zsh code.
Begin using common Zsh code. We implement a compiled command that will
catenate all the files it would otherwise source, so we can still have
a single file when we deploy `barnyard` by copying it to the target.

We are going to have new heredocs, a new argument parser, and we'll be
adding a library that makes it easier to illustrate the steps being
performed in module.